refactor feature gated code into submobules
