+++ src/py_try_from/mod.rs
@@ -16,7 +16,6 @@
 
 use std::collections::{BTreeMap, BTreeSet, HashMap, HashSet};
 
-use pyo3::types::PyComplex;
 use pyo3::{
     exceptions::PyValueError,
     types::{PyDate, PyDateTime, PyDelta, PyTime, PyTzInfo},
@@ -29,15 +28,8 @@ use pyo3::{
 };
 
 #[cfg(feature = "complex")]
-use num_complex::Complex;
-#[cfg(feature = "complex")]
-use num_traits::{Float, FloatConst};
-#[cfg(feature = "complex")]
-use pyo3::exceptions::PyFloatingPointError;
-#[cfg(feature = "complex")]
-use std::fmt::Display;
-#[cfg(feature = "complex")]
-use std::os::raw::c_double;
+/// Conversion trait implementations for complex numbers.
+mod complex;
 
 #[cfg(feature = "time")]
 use crate::datetime::DateTime;
@@ -47,7 +39,8 @@ use pyo3::{types::PyTuple, ToPyObject};
 use time::{Date, Duration, Month, OffsetDateTime, PrimitiveDateTime, Time, UtcOffset};
 
 #[cfg(feature = "indexmap")]
-use indexmap::IndexMap;
+/// Conversion trait implementations for indexmap.
+mod indexmap;
 
 /// Convert from a Python type to a Rust type.
 pub trait PyTryFrom<P>: Sized {
@@ -152,6 +145,7 @@ macro_rules! private_impl_py_try_from {
         }
     };
 }
+pub(crate) use private_impl_py_try_from;
 
 /// Implement [`PyTryFrom<PyAny>`] for a given Rust type by delegating to its implementation
 /// for the given Python type.
@@ -162,28 +156,31 @@ macro_rules! private_impl_py_try_from_with_pyany {
         $crate::private_impl_py_try_from_pyany!($py_type => $rs_type);
     };
 }
+pub(crate) use private_impl_py_try_from_with_pyany;
 
 /// Implements [`PyTryFrom<Py<P>>`] for `T` where `T: PyTryFrom<P>` and `P` is a native Python type.
 macro_rules! impl_try_from_py_native {
     ($py_type: ty => $rs_type: ty) => {
-        private_impl_py_try_from!(&item, py, $crate::pyo3::Py<$py_type> => $rs_type {
+        $crate::private_impl_py_try_from!(&item, py, $crate::pyo3::Py<$py_type> => $rs_type {
             let item: &$py_type = item.as_ref(py);
             <Self as $crate::PyTryFrom<$py_type>>::py_try_from(py, item)
         });
     }
 }
+pub(crate) use impl_try_from_py_native;
 
 /// Implements [`PyTryFrom<T>`] for a `T` that is a native Python type.
 macro_rules! impl_try_from_self_python {
     ($py_type: ty) => {
-        private_impl_py_try_from!(&item, py, $py_type => $crate::pyo3::Py<$py_type> {
+        $crate::private_impl_py_try_from!(&item, py, $py_type => $crate::pyo3::Py<$py_type> {
             Ok(item.into_py(py))
         });
-        private_impl_py_try_from!(&item, _py, $crate::pyo3::Py<$py_type> => $crate::pyo3::Py<$py_type> {
+        $crate::private_impl_py_try_from!(&item, _py, $crate::pyo3::Py<$py_type> => $crate::pyo3::Py<$py_type> {
             Ok(item.clone())
         });
     }
 }
+pub(crate) use impl_try_from_self_python;
 
 /// Implements [`PyTryFrom<T>`] for a `T` that is a native Rust type.
 macro_rules! impl_try_from_self_rust {
@@ -235,64 +232,6 @@ impl_try_from_py_native!(PyBytes => Box<[u8]>);
 private_impl_py_try_from!(&item, py, PyBytes => Box<[u8]> {
     Vec::<u8>::py_try_from(py, item).map(From::from)
 });
-
-// ==== Complex ====
-
-impl_try_from_self_python!(PyComplex);
-
-#[cfg(feature = "complex")]
-impl<F> PyTryFrom<Self> for Complex<F>
-where
-    F: Copy + Float + FloatConst + Into<c_double> + Display,
-{
-    fn py_try_from(_py: Python, item: &Self) -> PyResult<Self> {
-        Ok(*item)
-    }
-}
-
-#[cfg(feature = "complex")]
-impl<F> PyTryFrom<Py<PyComplex>> for Complex<F>
-where
-    F: Copy + Float + FloatConst + Into<c_double> + Display,
-{
-    fn py_try_from(py: Python, item: &Py<PyComplex>) -> PyResult<Self> {
-        Self::py_try_from(py, item.as_ref(py))
-    }
-}
-
-#[cfg(feature = "complex")]
-impl<F> PyTryFrom<PyComplex> for Complex<F>
-where
-    // `Display` seems like an odd trait to require, but it is used to make a more useful
-    // error message. The types realistically used for this are `f32` and `f64` both of which
-    // impl `Display`, so there's no issue there.
-    F: Copy + Float + FloatConst + Into<c_double> + Display,
-{
-    fn py_try_from(_py: Python, item: &PyComplex) -> PyResult<Self> {
-        let make_error = |val: c_double| {
-            PyFloatingPointError::new_err(format!(
-                "expected {val} to be between {} and {}, inclusive",
-                F::min_value(),
-                F::max_value(),
-            ))
-        };
-        Ok(Self {
-            re: F::from(item.real()).ok_or_else(|| make_error(item.real()))?,
-            im: F::from(item.imag()).ok_or_else(|| make_error(item.imag()))?,
-        })
-    }
-}
-
-#[cfg(feature = "complex")]
-impl<F> PyTryFrom<PyAny> for Complex<F>
-where
-    F: Copy + Float + FloatConst + Into<c_double> + Display,
-{
-    fn py_try_from(py: Python, item: &PyAny) -> PyResult<Self> {
-        let dict: &PyComplex = item.downcast()?;
-        Self::py_try_from(py, dict)
-    }
-}
 
 // ==== Date ====
 
@@ -828,68 +767,5 @@ private_impl_py_try_from_with_pyany!(&item, py, PyTzInfo => UtcOffset {
     })?;
     Ok(offset)
 });
-
-// ==== IndexMap ====
-
-#[cfg(feature = "indexmap")]
-impl<K1, K2, V1, V2, S> PyTryFrom<IndexMap<K1, V1, S>> for IndexMap<K2, V2, S>
-where
-    K2: Eq + std::hash::Hash + PyTryFrom<K1>,
-    V2: PyTryFrom<V1>,
-    S: std::hash::BuildHasher + Default,
-{
-    fn py_try_from(py: Python, item: &IndexMap<K1, V1, S>) -> PyResult<Self> {
-        item.iter()
-            .map(|(key, val)| {
-                let key = K2::py_try_from(py, key)?;
-                let val = V2::py_try_from(py, val)?;
-                Ok((key, val))
-            })
-            .collect()
-    }
-}
-
-#[cfg(feature = "indexmap")]
-impl<K, V, S> PyTryFrom<PyDict> for IndexMap<K, V, S>
-where
-    K: Eq + std::hash::Hash + PyTryFrom<PyAny>,
-    V: PyTryFrom<PyAny>,
-    S: std::hash::BuildHasher + Default,
-{
-    fn py_try_from(py: Python, item: &PyDict) -> PyResult<Self> {
-        let mut map = Self::with_capacity_and_hasher(item.len(), S::default());
-        for (key, val) in item {
-            let key = K::py_try_from(py, key)?;
-            let val = V::py_try_from(py, val)?;
-            map.insert(key, val);
-        }
-        Ok(map)
-    }
-}
-
-#[cfg(feature = "indexmap")]
-impl<K, V, S> PyTryFrom<Py<PyDict>> for IndexMap<K, V, S>
-where
-    K: Eq + std::hash::Hash + PyTryFrom<PyAny>,
-    V: PyTryFrom<PyAny>,
-    S: std::hash::BuildHasher + Default,
-{
-    fn py_try_from(py: Python, item: &Py<PyDict>) -> PyResult<Self> {
-        Self::py_try_from(py, item.as_ref(py))
-    }
-}
-
-#[cfg(feature = "indexmap")]
-impl<K, V, S> PyTryFrom<PyAny> for IndexMap<K, V, S>
-where
-    K: Eq + std::hash::Hash + PyTryFrom<PyAny>,
-    V: PyTryFrom<PyAny>,
-    S: std::hash::BuildHasher + Default,
-{
-    fn py_try_from(py: Python, item: &PyAny) -> PyResult<Self> {
-        let dict: &PyDict = item.downcast()?;
-        Self::py_try_from(py, dict)
-    }
-}
 
 // ============ End Implementations ==============
